Refactor all Task related files into chatty/task package
