+++ src/main/java/chatty/task/Deadline.java
@@ -1,4 +1,4 @@
-package chatty;
+package chatty.task;
 
 public class Deadline extends Task {
