+++ text-ui-test/runtest.sh
@@ -13,7 +13,7 @@ then
 fi
 
 # compile the code into the bin folder, terminates if error occurred
-if ! javac -cp ../src -Xlint:none -d ../bin ../src/main/java/chatty/*
+if ! javac -cp ../src -Xlint:none -d ../bin ../src/main/java/chatty/*.java ../src/main/java/chatty/task/*.java
 then
     echo "********** BUILD FAILURE **********"
     exit 1
